--- src/main/java/domain/CardsGenerator.java
@@ -1,9 +1,0 @@
-package domain;
-
-import domain.card.Card;
-import java.util.List;
-
-public interface CardsGenerator {
-
-    List<Card> generate();
-}
