+++ src/main/java/domain/card/ShuffledCardsGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import domain.card.Card;
 import domain.card.Rank;
