+++ src/main/java/domain/gamer/Gamers.java
@@ -1,8 +1,5 @@
-package domain;
+package domain.gamer;
 
-import domain.gamer.Dealer;
-import domain.gamer.Gamer;
-import domain.gamer.Player;
 import java.util.ArrayList;
 import java.util.List;
