+++ src/main/java/domain/result/PlayerResults.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 import domain.gamer.Player;
 import java.util.Collections;
