+++ src/main/java/domain/result/Result.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 import domain.gamer.Dealer;
 import domain.gamer.Player;
