+++ src/test/java/domain/card/DeckTest.java
@@ -1,8 +1,9 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.card.Card;
+import domain.card.Deck;
 import domain.card.Rank;
 import domain.card.Symbol;
 import java.util.List;
