+++ src/test/java/domain/card/ShuffledCardsGeneratorTest.java
@@ -1,8 +1,9 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import domain.card.Card;
+import domain.card.ShuffledCardsGenerator;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
