+++ src/test/java/domain/gamer/DealerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.gamer;
 
 import static org.assertj.core.api.Assertions.assertThat;
