+++ src/test/java/domain/gamer/PlayersTest.java
@@ -1,9 +1,10 @@
-package domain;
+package domain.gamer;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import domain.gamer.Name;
 import domain.gamer.Player;
+import domain.gamer.Players;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
